linux_wayland_support: dup detecting function of x11 or wayland
