+++ libs/enigo/Cargo.toml
@@ -22,6 +22,7 @@ appveyor = { repository = "pythoneer/enigo-85xiy" }
 serde = { version = "1.0", optional = true }
 serde_derive = { version = "1.0", optional = true }
 log = "0.4"
+hbb_common = { path = "../hbb_common" }
 
 [features]
 with_serde = ["serde", "serde_derive"]
