+++ libs/hbb_common/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod compress;
 pub mod protos;
+pub mod platform;
 pub use  protos::message as message_proto;
 pub use  protos::rendezvous as rendezvous_proto;
 pub use bytes;
